% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNetf-LibCombinded02.R,
%   R/standardStandaloneFunctions.R
\name{rnetf rnetf}
\alias{rnetf rnetf}
\alias{rnetf rnetf}
\title{A ml modular framework}
\arguments{
\item{rnetf}{type stuff}

\item{rnetf}{type stuff}
}
\description{
This function allows you to express your love of cats.

This function allows you to express your love of cats.
}
\examples{
rnetf()
nCompute()
nTrain()
}
\keyword{net}
\keyword{neural}
\keyword{neuralnetwork}
