% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNetf-LibCombinded02.R
\name{rnetf}
\alias{rnetf}
\title{A ml modular framework}
\arguments{
\item{rnetf}{type stuff}
}
\description{
This function allows you to express your love of cats.
}
\examples{
rnetf()
}
\keyword{neuralnetwork}
